+++ src/resources/index.ts
@@ -21,7 +21,7 @@ export {
   type CompanyGetContactV0Response,
   type CompanyGetDetailsV1Response,
   type CompanyGetFinancialsV1Response,
-  type CompanyGetHistoricalOwnersV0Response,
+  type CompanyGetHistoricalOwnersV1Response,
   type CompanyGetHoldingsV1Response,
   type CompanyGetOwnersV1Response,
   type CompanyGetUbosV1Response,